Move useSelectNearestEditableBlock out of src/hooks
